% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_assoc_metrics.R
\name{calc_assoc_metrics}
\alias{calc_assoc_metrics}
\title{Calculate Association Metrics for Bigrams}
\usage{
calc_assoc_metrics(
  data,
  doc_index,
  token_index,
  type,
  association = "all",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the corpus.}

\item{doc_index}{A string name of the column in 'data' which represents the document index.}

\item{token_index}{A string name of the column in 'data' which represents the token index.}

\item{type}{A string name of the column in 'data' which represents the tokens or terms.}

\item{association}{A character vector specifying which metrics to calculate. Can be any combination of 'pmi' (Pointwise Mutual Information), 'dice_coeff' (Dice's Coefficient), 'g_score' (G-score), or 'all' (calculate all metrics). Default is 'all'.}

\item{verbose}{A logical value indicating whether to keep the intermediate probability columns ('p_xy', 'p_x', 'p_y') in the result. Default is FALSE.}
}
\value{
A data frame with one row per bigram and columns for each calculated metric.
If 'verbose' is TRUE, the intermediate probabilities used in the calculations are also included.
}
\description{
This function calculates various association metrics (PMI, Dice's Coefficient, Lambda-Rank)
for bigrams in a given corpus. The data frame must contain document and token indices,
as well as a 'type' variable representing the tokens.
}
\examples{
\dontrun{
library(dplyr)
data <- tibble::tibble(
  doc_index = c(1, 1, 1, 2),
  token_index = c(1, 2, 3, 1),
  type = c("word1", "word2", "word3", "word2")
)
calc_assoc_metrics(data, doc_index, token_index, type)
}

}
