% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_dictionary.R
\name{create_data_dictionary}
\alias{create_data_dictionary}
\title{Create a Data Dictionary for a Given Dataset}
\usage{
create_data_dictionary(
  data,
  file_path,
  sample_n = 5,
  grouping = NULL,
  model = "gpt-3.5-turbo"
)
}
\arguments{
\item{data}{A data frame to create a data dictionary for.}

\item{file_path}{A character string specifying the file path to write the data dictionary to.}

\item{sample_n}{An integer specifying the number of rows to sample from the data frame.}

\item{grouping}{A character vector specifying the grouping variables to use when sampling the data frame.}

\item{model}{A character vector specifying the OpenAI chat completion model to use for generating the variable descriptions. Defaults to "gpt-3.5-turbo". See \code{\link[openai]{create_chat_completion}} for more information.}
}
\value{
A data frame containing the variable name, human-readable name, description, and variable type.
}
\description{
The data dictionary returned includes the variable name, human-readable name, description, and variable type. The function uses OpenAI's chat completions models to generate the descriptions for each of the variables in the dataset. The data dictionary is returned as a data frame and written to a file in CSV format.
}
\examples{
\dontrun{
create_data_dictionary(mtcars, "mtcars_data_dictionary.csv")
}
}
