% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_origin.R
\name{create_data_origin}
\alias{create_data_origin}
\title{Create data origin file}
\usage{
create_data_origin(file_path)
}
\arguments{
\item{file_path}{A character string specifying the file path where the data origin file should be saved.}
}
\value{
A tibble containing the data origin information.
}
\description{
This function creates a data frame with attributes about the origin of the data,
writes it to a CSV file at the specified file path, and returns the data frame.
}
\examples{
\dontrun{
create_data_origin("data_origin.csv")
}

}
