% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_talkbank_data.R
\name{get_talkbank_data}
\alias{get_talkbank_data}
\title{Downloads TalkBank data and saves it to disk}
\usage{
get_talkbank_data(
  corpus_name,
  corpus_path,
  target_dir,
  force = FALSE,
  confirmed = FALSE
)
}
\arguments{
\item{corpus_name}{The name of the TalkBank corpus to download data from.}

\item{corpus_path}{The path to the TalkBank corpus to download data from.}

\item{target_dir}{The directory to save the downloaded data to.}

\item{force}{If \code{TRUE}, the data will be downloaded even if it already exists on disk.}

\item{confirmed}{If \code{TRUE}, the user has confirmed that they have permission to use the data.
If \code{FALSE}, the function will prompt the user to confirm permission.
Setting this to \code{TRUE} is useful for reproducible workflows.}
}
\value{
A message indicating whether the data was acquired or already existed on disk.
}
\description{
Downloads the utterances, transcripts, participants, tokens, and token types data
from the TalkBank database and saves it to disk in the specified target directory.
}
\examples{
\dontrun{
# Download CABNC data from the Conversation Bank to a directory called "data"
get_talkbank_data(corpus_name = "ca", corpus_path = c("ca", "CABNC"), target_dir = "data")
}
}
