% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtalrkit-package.R
\docType{package}
\name{qtalrkit-package}
\alias{qtalrkit}
\alias{qtalrkit-package}
\title{qtalrkit: Quantitative Text Analysis for Linguists Resource Kit}
\description{
Support package for the textbook "An Introduction to Quantitative Text Analysis for Linguists: Reproducible Research using R". Includes access to interactive code exercises and demos, data, and misc functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qtalr/qtalrkit}
  \item \url{https://qtalr.github.io/qtalrkit/}
  \item Report bugs at \url{https://github.com/qtalr/qtalrkit/issues}
}

}
\author{
\strong{Maintainer}: Jerid Francom \email{francojc@wfu.edu} (\href{https://orcid.org/0000-0001-5972-6330}{ORCID})

}
\keyword{internal}
